+++ lib/provider/timeline_last_viewed_note_id_notifier_provider.dart
@@ -4,7 +4,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../model/id.dart';
 import '../model/tab_settings.dart';
-import 'api/id_gen_method_provider.dart';
+import 'id_gen_method_provider.dart';
 import 'shared_preferences_provider.dart';
 
 part 'timeline_last_viewed_note_id_notifier_provider.g.dart';
