+++ lib/view/page/search/search_notes.dart
@@ -9,9 +9,9 @@ import '../../../i18n/strings.g.dart';
 import '../../../model/account.dart';
 import '../../../model/id.dart';
 import '../../../provider/api/channel_notifier_provider.dart';
-import '../../../provider/api/id_gen_method_provider.dart';
 import '../../../provider/api/search_notes_notifier_provider.dart';
 import '../../../provider/api/user_notifier_provider.dart';
+import '../../../provider/id_gen_method_provider.dart';
 import '../../../util/format_datetime.dart';
 import '../../../util/pick_date_time.dart';
 import '../../../util/punycode.dart';
